refactor: move profile repository stuff to it own directory
