+++ src/app/contracts/repositories/profiles/i-get-profiles-repository.ts
@@ -1,4 +1,4 @@
-import { IWhereClauseOption } from "./common/i-where-clause-option";
+import { IWhereClauseOption } from "../common/i-where-clause-option";
 import { IProfileModel } from "./i-profile-model";
 
 export type IGetProfilesRepositoryOptions = {
